+++ lib/ain-contracts/build.rs
@@ -17,11 +17,11 @@ fn main() -> Result<()> {
         ("dfi_reserved", "DFIReserved"),
         ("dfi_intrinsics_registry", "DFIIntrinsicsRegistry"),
         ("dfi_intrinsics_v1", "DFIIntrinsicsV1"),
-        ("dfi_intrinsics_v2", "DFIIntrinsicsV2"),
         ("transfer_domain", "TransferDomain"),
         ("transfer_domain_v1", "TransferDomainV1"),
         ("dst20", "DST20"),
         ("dst20_v1", "DST20V1"),
+        ("dst20_v2", "DST20V2"),
     ];
 
     for (sol_project_name, contract_name) in contracts {
