--- lib/ain-contracts/dfi_intrinsics_v2/IDFIIntrinsicsV2.sol
@@ -1,11 +1,0 @@
-// SPDX-License-Identifier: MIT
-pragma solidity ^0.8.0;
-
-interface IDFIIntrinsicsV1 { 
-    function version() external view returns (uint256);
-    function evmBlockCount() external view returns(uint256);
-    function dvmBlockCount() external view returns(uint256);
-}
-
-interface IDFIIntrinsicsV2 is IDFIIntrinsicsV1 {
-}
