+++ lib/ain-contracts/dst20_v2/ITokenSplit.sol
@@ -1,0 +1,11 @@
+// SPDX-License-Identifier: MIT
+pragma solidity ^0.8.0;
+
+interface ITokenSplit {
+    event SplitResult(
+        address indexed newTokenContractAddress,
+        uint256 newAmount
+    );
+
+    function split(uint256 amount) external returns (address, uint256);
+}
