+++ lib/ain-evm/src/precompiles/mod.rs
@@ -7,7 +7,7 @@ mod token_split;
 #[cfg(test)]
 mod test_vector_support;
 
-use ain_contracts::get_dfi_intrinsics_v2_contract;
+use ain_contracts::DST20_ADDR_PREFIX_BYTE;
 use blake2::Blake2F;
 use bn128::{Bn128Add, Bn128Mul, Bn128Pairing};
 use ethereum_types::H160;
@@ -116,9 +116,6 @@ impl MetachainPrecompiles {
 
 impl PrecompileSet for MetachainPrecompiles {
     fn execute(&self, handle: &mut impl PrecompileHandle) -> Option<PrecompileResult> {
-        let is_intrinsics =
-            || handle.context().caller == get_dfi_intrinsics_v2_contract().fixed_address;
-
         match handle.code_address() {
             a if a == hash(1) => Some(<ECRecover as Precompile>::execute(handle)),
             a if a == hash(2) => Some(<Sha256 as Precompile>::execute(handle)),
@@ -130,7 +127,7 @@ impl PrecompileSet for MetachainPrecompiles {
             a if a == hash(8) => Some(Bn128Pairing::execute(handle)),
             a if a == hash(9) => Some(Blake2F::execute(handle)),
 
-            a if a == hash(10) && is_intrinsics() => {
+            a if a == hash(10) && is_dst20(handle.context().caller) => {
                 let mnview_ptr = self.0.unwrap_or_default(); // If None, should fetch from global view
                 Some(TokenSplit::execute(handle, mnview_ptr))
             }
@@ -148,4 +145,15 @@ impl PrecompileSet for MetachainPrecompiles {
 
 fn hash(a: u64) -> H160 {
     H160::from_low_u64_be(a)
+}
+
+fn is_dst20(addr: H160) -> bool {
+    match addr.as_fixed_bytes() {
+        [prefix, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, ..]
+            if prefix == &DST20_ADDR_PREFIX_BYTE =>
+        {
+            true
+        }
+        _ => false,
+    }
 }
